Add files via upload
This is the bug report cog, this as of now, will take the command, will send user a DM asking the the muser to send their bug report along with discription and and a screen hshot if possible. right now , the cog does not send the bug report to the bug-report channel.