Add files via upload
Bug report works but does not send the bug info from user to specified bug-report channel. (still figuring it out)
Flipcoin takes user input and compares it to a random integer between 1 and 2, representing heads and tails.
The pomodoro timer asks the user to focus on one task for 25 minutes and then informs the user to take a break for five minutess. It does this 16 times, so basically for the whole 8 hour shif.


The reminder was updated so that it doesnt take anything but an integer as input from the user for reminder time in minutes